Разбил репозиторий на три сущности: GitHubUsersRepository (пользователи), RepositoriesGitHubUserRepository (репозитории пользователей) и GitHubImageRepository (изображение Github).
