+++ app/src/main/java/com/example/githubusers/GitHubUsersApp.kt
@@ -1,18 +1,30 @@
 package com.example.githubusers
 
 import android.app.Application
-import com.example.githubusers.data.repository.GitHubRepositoryImpl
+import com.example.githubusers.data.repository.GitHubImageRepositoryImpl
+import com.example.githubusers.data.repository.GitHubUsersRepositoryImpl
+import com.example.githubusers.data.repository.RepositoriesGitHubUserRepositoryImpl
 import com.example.githubusers.data.repository.network.NetWorkStatus
 import com.example.githubusers.data.room.DatabaseGitHubUsers
 import com.example.githubusers.di.AppComponent
 import com.example.githubusers.di.DaggerAppComponent
 import com.example.githubusers.di.modules.AppModule
-import com.example.githubusers.domain.repository.IGitHubRepository
+import com.example.githubusers.domain.repository.IGitHubImageRepository
+import com.example.githubusers.domain.repository.IGitHubUsersRepository
+import com.example.githubusers.domain.repository.IRepositoriesGitHubUserRepository
 
 class GitHubUsersApp : Application() {
 
-    val gitHubRepository: IGitHubRepository by lazy {
-        GitHubRepositoryImpl(NetWorkStatus(this))
+    val gitHubUsersRepository: IGitHubUsersRepository by lazy {
+        GitHubUsersRepositoryImpl(NetWorkStatus(this))
+    }
+
+    val repositoriesGitHubUserRepository: IRepositoriesGitHubUserRepository by lazy {
+        RepositoriesGitHubUserRepositoryImpl(NetWorkStatus(this))
+    }
+
+    val gitHubImageRepository: IGitHubImageRepository by lazy {
+        GitHubImageRepositoryImpl()
     }
 
     lateinit var appComponent: AppComponent
