+++ app/src/main/java/com/example/githubusers/domain/repository/IGitHubUsersRepository.kt
@@ -1,0 +1,10 @@
+package com.example.githubusers.domain.repository
+
+import com.example.githubusers.domain.dto.GitHubUser
+import io.reactivex.rxjava3.core.Observable
+import io.reactivex.rxjava3.core.Single
+
+interface IGitHubUsersRepository {
+    fun getGitHubUsers(): Single<List<GitHubUser>>
+    fun getDefaultGitHubUsers(): Observable<List<GitHubUser>>
+}
