+++ app/src/main/java/com/example/githubusers/ui/git_hub_users/GitHubUsersActivity.kt
@@ -34,7 +34,9 @@ class GitHubUsersActivity : MvpAppCompatActivity(), IGitHubUsersView {
     private lateinit var binding: ActivityGitHubUsersBinding
     private val gitHubUsersPresenter by moxyPresenter {
         GitHubUsersPresenter(
-            GitHubUsersApp.instance.gitHubRepository,
+            GitHubUsersApp.instance.gitHubUsersRepository,
+            GitHubUsersApp.instance.repositoriesGitHubUserRepository,
+            GitHubUsersApp.instance.gitHubImageRepository,
             AndroidSchedulers.mainThread()
         ).apply {
             GitHubUsersApp.instance.appComponent.inject(this)
